Completed Removing Duplicate actions button #1076
