+++ app/next-client-app/app/(protected)/scanreports/[id]/mapping_rules/rules-buttons.tsx
@@ -2,23 +2,9 @@
 
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogTrigger } from "@/components/ui/dialog";
-import {
-  BarChartHorizontalBig,
-  ChevronDown,
-  FileJson,
-  FileSpreadsheet,
-} from "lucide-react";
+import { BarChartHorizontalBig } from "lucide-react";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { GetFile } from "@/app/(protected)/scanreports/[id]/mapping_rules/get-file";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { requestFile } from "@/api/files";
-import { toast } from "sonner";
-import { useRouter } from "next/navigation";
 
 export function RulesButton({
   scanreportId,
@@ -29,19 +15,7 @@ export function RulesButton({
   query: string;
   filename: string;
 }) {
-  const router = useRouter();
-
-  const handleDownload = async (fileType: FileTypeFormat) => {
-    const resp = await requestFile(Number(scanreportId), fileType);
-    if (resp.success) {
-      router.push(`downloads`);
-      toast.success("File requested.");
-    } else {
-      toast.error(
-        `Error downloading file: ${(resp.errorMessage as any).message}`
-      );
-    }
-  };
+ 
   return (
     <div className="hidden md:flex gap-2 justify-end w-full mr-2">
       <div>
@@ -65,37 +39,6 @@ export function RulesButton({
             </ScrollArea>
           </DialogContent>
         </Dialog>
-      </div>
-      <div>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="outline">
-              Request Download <ChevronDown className="ml-2 size-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent className="w-[180px]">
-            <DropdownMenuItem>
-              <Button
-                onClick={() => handleDownload("application/json")}
-                variant={"ghost"}
-                size={"sm"}
-              >
-                Mapping JSON
-                <FileJson className="ml-2 size-4" />
-              </Button>
-            </DropdownMenuItem>
-            <DropdownMenuItem>
-              <Button
-                onClick={() => handleDownload("text/csv")}
-                variant={"ghost"}
-                size={"sm"}
-              >
-                Mapping CSV
-                <FileSpreadsheet className="ml-2 size-4" />
-              </Button>
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
       </div>
     </div>
   );
